break

+++ backend/controller/userController/controller.go
@@ -3,7 +3,6 @@ package usercontroller
 import (
 	"backend/controller"
 	"backend/model"
-	"backend/pkg/logjson"
 	"fmt"
 	"mime/multipart"
 	"net/http"
@@ -37,7 +36,6 @@ type UpdateProfileRequest struct {
 func UpdateProfile(c *gin.Context) {
 	user := c.MustGet("user").(*model.User)
 
-	fmt.Println(c.GetHeader("Content-Type"))
 	var userUpdate UpdateProfileRequest
 	err := c.ShouldBindWith(&userUpdate, binding.FormMultipart)
 	if err != nil {
@@ -49,13 +47,13 @@ func UpdateProfile(c *gin.Context) {
 	user.Nama = userUpdate.Nama
 	fileSavedPath := path.Join("storage", "image", "profile", fmt.Sprintf("%d%d.png", time.Now().Unix(), user.ID))
 
-	fmt.Println(fileSavedPath)
-	logjson.ToJSON(userUpdate)
 	if userUpdate.Avatar != nil {
-		removeErr := os.Remove(*user.Avatar)
-		if removeErr != nil {
-			c.JSON(http.StatusInternalServerError, controller.Response(controller.Error, removeErr.Error(), nil))
-			return
+		if user.Avatar != nil {
+			removeErr := os.Remove(*user.Avatar)
+			if removeErr != nil {
+				c.JSON(http.StatusInternalServerError, controller.Response(controller.Error, removeErr.Error(), nil))
+				return
+			}
 		}
 
 		if err := c.SaveUploadedFile(userUpdate.Avatar, fileSavedPath); err != nil {
@@ -65,7 +63,6 @@ func UpdateProfile(c *gin.Context) {
 		user.Avatar = &fileSavedPath
 	}
 
-	logjson.ToJSON(user)
 	Updateerr := model.UpdateUser(user)
 	if Updateerr != nil {
 		c.JSON(http.StatusInternalServerError, controller.Response(controller.Error, Updateerr.Error(), nil))
@@ -80,7 +77,6 @@ func AddPesanan(c *gin.Context) {
 	var formData PesananRequest
 	err := c.ShouldBindJSON(&formData)
 	if err != nil {
-		fmt.Println(err)
 		c.JSON(http.StatusBadRequest, controller.Response(controller.Error, err.Error(), nil))
 		return
 	}
